ai_services/services/virtual_tryon_service.py
import cv2
import numpy as np
from typing import Dict, Any
import io
from PIL import Image
import asyncio

class VirtualTryOnService:
    def __init__(self):
        self.initialized = True

    async def generate_tryon(
        self,
        user_image_url: str,
        product_image_url: str,
        body_landmarks: list,
        product_metadata: Dict[str, Any]
    ) -> Dict[str, Any]:
        """Generate virtual try-on result"""
        try:
            # Simulate processing time
            await asyncio.sleep(2)
            
            # In a real implementation, this would:
            # 1. Load user image and product image
            # 2. Use body landmarks to position clothing
            # 3. Apply realistic lighting and shadows
            # 4. Generate final composite image
            
            # Mock result for now
            result = {
                "result_image_url": "https://example.com/tryon-result.jpg",
                "fit_score": 0.87,
                "processing_time_ms": 2000,
                "confidence": 0.92,
                "recommendations": {
                    "fit_feedback": "Great fit! The size looks perfect for your body type.",
                    "style_notes": "This color complements your skin tone well.",
                    "alternative_sizes": []
                }
            }
            
            return result
            
        except Exception as e:
            raise Exception(f"Virtual try-on generation failed: {str(e)}")

    def _load_image_from_url(self, url: str) -> np.ndarray:
        """Load image from URL"""
        # Implementation would download and load image
        pass

    def _apply_clothing_overlay(
        self,
        user_image: np.ndarray,
        clothing_image: np.ndarray,
        landmarks: list
    ) -> np.ndarray:
        """Apply clothing overlay using body landmarks"""
        # Implementation would use computer vision techniques
        # to realistically overlay clothing on the user
        pass

    def _calculate_fit_score(
        self,
        body_measurements: Dict[str, float],
        clothing_measurements: Dict[str, float]
    ) -> float:
        """Calculate how well the clothing fits"""
        # Implementation would compare measurements
        # and return a fit score between 0 and 1
        return 0.87