ai_services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
opencv-python==4.8.1.78
mediapipe==0.10.8
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
boto3==1.34.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0